pipeline/requirements.txt
yfinance
polars
pyarrow
tqdm
pytest
matplotlib
feedparser
mlflow
google-genai